server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv==0.21.1
pyyaml==6.0.1
requests==2.31.0
openai==1.3.7
python-dateutil==2.8.2
websocket-client==1.5.1
edge-tts==7.0.0
moviepy==1.0.3
numpy>=1.21.6
imageio-ffmpeg>=0.2.0
proglog<=1.0.0
